Clean code

+++ convoexpress-frontend/src/components/AccountSettings.jsx
@@ -4,16 +4,19 @@ import axios from 'axios'
 
 const AccountSettings = () => {
 
-    const {setId,setUsername, setWs, setShowUpdateForm} = useGlobalContext();
+    const {username, id,setId,setUsername, setWs} = useGlobalContext();
 
     const deleteAccount = async (userId) => {
-        try{
-            await axios.delete(`/api/delete/${userId}`);
-            setWs(null)
-            setId(null);
-            setUsername(null);
-        }catch(err){
-            console.log('error',err);
+        const response = confirm(`Are you sure you to delete this ${username} account?`);
+        if(response){
+            try{
+                await axios.delete(`/api/delete/${userId}`);
+                setWs(null)
+                setId(null);
+                setUsername(null);
+            }catch(err){
+                console.log('error',err);
+            }
         }
     };
 
@@ -29,11 +32,10 @@ const AccountSettings = () => {
     };
 
   return (
-    <div className='p-4 text-sm flex flex-col items-center gap-4'>
+    <div className='p-4 text-sm flex flex-col items-center gap-4 border-t border-gray-700'>
         <p className='text-text-fade'>Account Settings</p>
         <div className='w-full flex justify-between underline text-accent'>
             <button className='hover:text-white' onClick={() => deleteAccount(id)}>Delete</button>
-            <button className='hover:text-white' onClick={() => setShowUpdateForm(prev => !prev)}>Update</button>
             <button className='hover:text-white' onClick={logoutUser}>Logout</button>
         </div>
     </div>
